// src/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        UniqueUser: {
            /** @example asdjas0sd-sdasdi-asdasd */
            id?: string;
            /** @example tester */
            username?: string;
            /** @example Tester */
            name?: string;
            /** @example albert@einstein.edu */
            mailAddress?: string;
        };
        User: {
            /** @example tester */
            username: string;
            /** @example Tester */
            name: string;
            /** @example albert@einstein.edu */
            mailAddress: string;
        };
        FullUser: {
            /** @example Tester */
            username: string;
            /**
             * Format: password
             * @example jshHfu3ek=fDÖ
             */
            password: string;
            /** @example Tester */
            name?: string;
            /** @example albert@einstein.edu */
            mailAddress?: string;
        };
// ...
    };
    responses: {
        /** @description Access token is missing or invalid; try to refresh your token */
        UnauthorizedError: {
            headers: {
                [name: string]: unknown;
            };
            content?: never;
        };
    };
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}